выполняет задание Модуляция
