+++ index.html
@@ -8,7 +8,7 @@
   <link rel="stylesheet" href="css/style.css">
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
   <script src="/js/functions.js" defer></script>
-  <script src="/js/main.js" defer></script>
+  <script src="/js/main.js" type="module" defer></script>
   <title>Кекстаграм</title>
 </head>
